requirements.txt
django-filter==2.2
django == 2.2
djangorestframework==3.11.1
djangorestframework-filters